Cha10 - Update History example code
In terms of performance, we could discuss if the `if` statement is better or not before adding an element to a set..

```
➤ ipython
Python 3.6.8 (default, Jan 14 2019, 11:02:34) 
Type 'copyright', 'credits' or 'license' for more information
IPython 7.13.0 -- An enhanced Interactive Python. Type '?' for help.

In [1]: s = set()                                                                                                                                                                                                  

In [2]: %timeit s.add("a")                                                                                                                                                                                         
41.8 ns ± 0.447 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [3]: s = set()                                                                                                                                                                                                  

In [4]: %timeit if not "a" in s: s.add("a")                                                                                                                                                                        
19.4 ns ± 0.102 ns per loop (mean ± std. dev. of 7 runs, 100000000 loops each)
```

In my point of view, as we're using a `set` for `HistoryEntry`, checking presence prior adding it adds a bit of noise in your example The shorter it is, the quicker it is to reasonate about...

But I guess one could think otherwise. Let me know !